refactor: extract webhook.Handler from webhook.Server
Step 4b-1 of the webhook split (issue #250). Establishes a dedicated
Handler type for the GitHub-webhook receiver in internal/webhook so the
package boundary becomes obvious before the larger lifecycle move.

Architecture:
  - New webhook.Handler holds the GitHub-webhook receiver state
    (delivery, channels, cfg, logger) and owns:
      - HMAC SHA-256 signature verification
      - Delivery dedupe via the existing DeliveryStore
      - The 6 GitHub event-type parsers (issues, pull_request,
        issue_comment, pull_request_review,
        pull_request_review_comment, push)
      - The top-level handleGitHubWebhook entry point
      - RegisterRoutes (mounts POST {cfg.Daemon.HTTP.WebhookPath})
  - webhook.Server keeps a *Handler field constructed in NewServer.
    Server.handleGitHubWebhook becomes a one-line delegate so existing
    tests that call it directly keep working without modification.
  - Handler.RegisterRoutes exists but isn't wired yet — buildHandler
    still mounts the route via Server.handleGitHubWebhook. That's
    deferred to 4b-2 when the lifecycle moves out of webhook.

webhook.Server shrinks ~408 lines (event handlers, payload types,
verifySignature, enqueue all moved). No behaviour change, no test
churn — every existing test passes without modification.
