+++ internal/webhook/server.go
@@ -2,16 +2,12 @@ package webhook
 
 import (
 	"context"
-	"crypto/hmac"
-	"crypto/sha256"
 	"database/sql"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"io"
 	"io/fs"
 	"net/http"
-	"strings"
 	"sync"
 	"time"
 
@@ -34,6 +30,7 @@ type Server struct {
 	delivery      *DeliveryStore
 	logger        zerolog.Logger
 	channels      server.EventQueue
+	h             *Handler // GitHub webhook receiver — owns event parsing, HMAC, dedupe
 	provider      server.StatusProvider
 	runtimeState  server.RuntimeStateProvider // optional; used by /api/agents for live run status
 	dispatchStats server.DispatchStatsProvider
@@ -175,6 +172,11 @@ func NewServer(cfg *config.Config, delivery *DeliveryStore, channels server.Even
 	// effective YAML-loaded snapshot before WithStore wires /export +
 	// /import.
 	s.config = serverconfig.New(s, s, s.logger)
+	// GitHub webhook receiver — pure event-parsing + HMAC + delivery dedupe.
+	// The server delegates handleGitHubWebhook to it so existing tests that
+	// call s.handleGitHubWebhook keep working unchanged; future PRs will
+	// drop that delegate and mount h.RegisterRoutes directly.
+	s.h = NewHandler(delivery, channels, s, logger)
 	if cfg.Daemon.Proxy.Enabled {
 		up := cfg.Daemon.Proxy.Upstream
 		s.proxy = anthropicproxy.NewHandler(anthropicproxy.UpstreamConfig{
@@ -459,410 +461,13 @@ func (s *Server) handleAgentsRun(w http.ResponseWriter, r *http.Request) {
 	})
 }
 
+// handleGitHubWebhook delegates to the package-internal Handler, which
+// owns HMAC verification, delivery dedupe, and per-event-type parsing.
+// Kept on the Server type so existing tests that call s.handleGitHubWebhook
+// directly continue to work; the route in buildHandler still goes through
+// here for the same reason. A follow-up PR will mount h.RegisterRoutes
+// directly and drop this method.
 func (s *Server) handleGitHubWebhook(w http.ResponseWriter, r *http.Request) {
-	cfg := s.loadCfg()
-	deliveryID := strings.TrimSpace(r.Header.Get("X-GitHub-Delivery"))
-	if deliveryID == "" {
-		http.Error(w, "missing delivery id", http.StatusBadRequest)
-		return
-	}
-
-	body, err := io.ReadAll(http.MaxBytesReader(w, r.Body, cfg.Daemon.HTTP.MaxBodyBytes))
-	if err != nil {
-		http.Error(w, "invalid body", http.StatusBadRequest)
-		return
-	}
-	if !verifySignature(body, cfg.Daemon.HTTP.WebhookSecret, r.Header.Get("X-Hub-Signature-256")) {
-		http.Error(w, "invalid signature", http.StatusUnauthorized)
-		return
-	}
-	// Delivery dedup is checked only after signature verification so
-	// unauthenticated requests cannot poison the dedupe cache.
-	if s.delivery.SeenOrAdd(deliveryID, time.Now()) {
-		w.WriteHeader(http.StatusAccepted)
-		return
-	}
-
-	event := strings.TrimSpace(r.Header.Get("X-GitHub-Event"))
-	switch event {
-	case "issues":
-		s.handleIssuesEvent(r.Context(), w, body, deliveryID)
-	case "pull_request":
-		s.handlePullRequestEvent(r.Context(), w, body, deliveryID)
-	case "issue_comment":
-		s.handleIssueCommentEvent(r.Context(), w, body, deliveryID)
-	case "pull_request_review":
-		s.handlePullRequestReviewEvent(r.Context(), w, body, deliveryID)
-	case "pull_request_review_comment":
-		s.handlePullRequestReviewCommentEvent(r.Context(), w, body, deliveryID)
-	case "push":
-		s.handlePushEvent(r.Context(), w, body, deliveryID)
-	default:
-		s.logger.Warn().Str("event", event).Str("delivery_id", deliveryID).Msg("unhandled webhook event type")
-		w.WriteHeader(http.StatusAccepted)
-	}
+	s.h.handleGitHubWebhook(w, r)
 }
 
-// ─── webhook payload shapes ───────────────────────────────────────────────────
-
-type webhookRepository struct {
-	FullName string `json:"full_name"`
-}
-
-type webhookSender struct {
-	Login string `json:"login"`
-}
-
-type webhookLabel struct {
-	Name string `json:"name"`
-}
-
-type webhookIssue struct {
-	Number      int       `json:"number"`
-	Title       string    `json:"title"`
-	Body        string    `json:"body"`
-	PullRequest *struct{} `json:"pull_request,omitempty"`
-}
-
-type webhookPullRequest struct {
-	Number int    `json:"number"`
-	Title  string `json:"title"`
-	Draft  bool   `json:"draft"`
-	Merged bool   `json:"merged"`
-}
-
-type webhookComment struct {
-	Body string `json:"body"`
-}
-
-type webhookReview struct {
-	Body  string `json:"body"`
-	State string `json:"state"`
-}
-
-// ─── event-type handlers ──────────────────────────────────────────────────────
-
-// handleIssuesEvent handles X-GitHub-Event: issues.
-// For "labeled" actions it filters to AI labels and emits "issues.labeled".
-// For lifecycle actions (opened, edited, reopened, closed) it emits the
-// corresponding "issues.{action}" event.
-// Events from issues that are pull requests (GitHub sends both) are dropped
-// for the "labeled" action; the pull_request event handles those.
-func (s *Server) handleIssuesEvent(ctx context.Context, w http.ResponseWriter, body []byte, deliveryID string) {
-	cfg := s.loadCfg()
-	var payload struct {
-		Action     string            `json:"action"`
-		Label      webhookLabel      `json:"label"`
-		Issue      webhookIssue      `json:"issue"`
-		Repository webhookRepository `json:"repository"`
-		Sender     webhookSender     `json:"sender"`
-	}
-	if err := json.Unmarshal(body, &payload); err != nil {
-		http.Error(w, "invalid payload", http.StatusBadRequest)
-		return
-	}
-
-	repo, ok := cfg.RepoByName(payload.Repository.FullName)
-	if !ok || !repo.Enabled {
-		w.WriteHeader(http.StatusAccepted)
-		return
-	}
-
-	repoRef := workflow.RepoRef{FullName: repo.Name, Enabled: repo.Enabled}
-
-	// GitHub sends issues events for PR-backed issues too; the pull_request event
-	// handles those, so drop all issue events that belong to a pull request.
-	if payload.Issue.PullRequest != nil {
-		w.WriteHeader(http.StatusAccepted)
-		return
-	}
-
-	switch payload.Action {
-	case "labeled":
-		ev := workflow.Event{
-			ID:     deliveryID,
-			Repo:   repoRef,
-			Kind:   "issues.labeled",
-			Number: payload.Issue.Number,
-			Actor:  payload.Sender.Login,
-			Payload: map[string]any{
-				"label": payload.Label.Name,
-			},
-		}
-		s.enqueue(ctx, w, ev, deliveryID)
-	case "opened", "edited", "reopened", "closed":
-		ev := workflow.Event{
-			ID:     deliveryID,
-			Repo:   repoRef,
-			Kind:   "issues." + payload.Action,
-			Number: payload.Issue.Number,
-			Actor:  payload.Sender.Login,
-			Payload: map[string]any{
-				"title": payload.Issue.Title,
-				"body":  payload.Issue.Body,
-			},
-		}
-		s.enqueue(ctx, w, ev, deliveryID)
-	default:
-		w.WriteHeader(http.StatusAccepted)
-	}
-}
-
-// handlePullRequestEvent handles X-GitHub-Event: pull_request.
-// For "labeled" actions it filters to AI labels (and skips drafts) and emits
-// "pull_request.labeled". For lifecycle actions it emits "pull_request.{action}".
-func (s *Server) handlePullRequestEvent(ctx context.Context, w http.ResponseWriter, body []byte, deliveryID string) {
-	cfg := s.loadCfg()
-	var payload struct {
-		Action      string             `json:"action"`
-		Label       webhookLabel       `json:"label"`
-		PullRequest webhookPullRequest `json:"pull_request"`
-		Repository  webhookRepository  `json:"repository"`
-		Sender      webhookSender      `json:"sender"`
-	}
-	if err := json.Unmarshal(body, &payload); err != nil {
-		http.Error(w, "invalid payload", http.StatusBadRequest)
-		return
-	}
-
-	repo, ok := cfg.RepoByName(payload.Repository.FullName)
-	if !ok || !repo.Enabled {
-		w.WriteHeader(http.StatusAccepted)
-		return
-	}
-
-	repoRef := workflow.RepoRef{FullName: repo.Name, Enabled: repo.Enabled}
-
-	switch payload.Action {
-	case "labeled":
-		if payload.PullRequest.Draft {
-			s.logger.Info().Str("repo", repo.Name).Int("number", payload.PullRequest.Number).Msg("pull request skipped, draft")
-			w.WriteHeader(http.StatusAccepted)
-			return
-		}
-		ev := workflow.Event{
-			ID:     deliveryID,
-			Repo:   repoRef,
-			Kind:   "pull_request.labeled",
-			Number: payload.PullRequest.Number,
-			Actor:  payload.Sender.Login,
-			Payload: map[string]any{
-				"label": payload.Label.Name,
-			},
-		}
-		s.enqueue(ctx, w, ev, deliveryID)
-	case "opened", "synchronize", "ready_for_review", "closed":
-		eventPayload := map[string]any{
-			"title": payload.PullRequest.Title,
-			"draft": payload.PullRequest.Draft,
-		}
-		if payload.Action == "closed" {
-			eventPayload["merged"] = payload.PullRequest.Merged
-		}
-		ev := workflow.Event{
-			ID:      deliveryID,
-			Repo:    repoRef,
-			Kind:    "pull_request." + payload.Action,
-			Number:  payload.PullRequest.Number,
-			Actor:   payload.Sender.Login,
-			Payload: eventPayload,
-		}
-		s.enqueue(ctx, w, ev, deliveryID)
-	default:
-		w.WriteHeader(http.StatusAccepted)
-	}
-}
-
-// handleIssueCommentEvent handles X-GitHub-Event: issue_comment.
-// Only "created" actions are forwarded as "issue_comment.created".
-func (s *Server) handleIssueCommentEvent(ctx context.Context, w http.ResponseWriter, body []byte, deliveryID string) {
-	cfg := s.loadCfg()
-	var payload struct {
-		Action     string            `json:"action"`
-		Comment    webhookComment    `json:"comment"`
-		Issue      webhookIssue      `json:"issue"`
-		Repository webhookRepository `json:"repository"`
-		Sender     webhookSender     `json:"sender"`
-	}
-	if err := json.Unmarshal(body, &payload); err != nil {
-		http.Error(w, "invalid payload", http.StatusBadRequest)
-		return
-	}
-	if payload.Action != "created" {
-		w.WriteHeader(http.StatusAccepted)
-		return
-	}
-
-	repo, ok := cfg.RepoByName(payload.Repository.FullName)
-	if !ok || !repo.Enabled {
-		w.WriteHeader(http.StatusAccepted)
-		return
-	}
-
-	ev := workflow.Event{
-		ID:     deliveryID,
-		Repo:   workflow.RepoRef{FullName: repo.Name, Enabled: repo.Enabled},
-		Kind:   "issue_comment.created",
-		Number: payload.Issue.Number,
-		Actor:  payload.Sender.Login,
-		Payload: map[string]any{
-			"body": payload.Comment.Body,
-		},
-	}
-	s.enqueue(ctx, w, ev, deliveryID)
-}
-
-// handlePullRequestReviewEvent handles X-GitHub-Event: pull_request_review.
-// Only "submitted" actions are forwarded as "pull_request_review.submitted".
-func (s *Server) handlePullRequestReviewEvent(ctx context.Context, w http.ResponseWriter, body []byte, deliveryID string) {
-	cfg := s.loadCfg()
-	var payload struct {
-		Action      string             `json:"action"`
-		Review      webhookReview      `json:"review"`
-		PullRequest webhookPullRequest `json:"pull_request"`
-		Repository  webhookRepository  `json:"repository"`
-		Sender      webhookSender      `json:"sender"`
-	}
-	if err := json.Unmarshal(body, &payload); err != nil {
-		http.Error(w, "invalid payload", http.StatusBadRequest)
-		return
-	}
-	if payload.Action != "submitted" {
-		w.WriteHeader(http.StatusAccepted)
-		return
-	}
-
-	repo, ok := cfg.RepoByName(payload.Repository.FullName)
-	if !ok || !repo.Enabled {
-		w.WriteHeader(http.StatusAccepted)
-		return
-	}
-
-	ev := workflow.Event{
-		ID:     deliveryID,
-		Repo:   workflow.RepoRef{FullName: repo.Name, Enabled: repo.Enabled},
-		Kind:   "pull_request_review.submitted",
-		Number: payload.PullRequest.Number,
-		Actor:  payload.Sender.Login,
-		Payload: map[string]any{
-			"state": payload.Review.State,
-			"body":  payload.Review.Body,
-		},
-	}
-	s.enqueue(ctx, w, ev, deliveryID)
-}
-
-// handlePullRequestReviewCommentEvent handles X-GitHub-Event: pull_request_review_comment.
-// Only "created" actions are forwarded as "pull_request_review_comment.created".
-func (s *Server) handlePullRequestReviewCommentEvent(ctx context.Context, w http.ResponseWriter, body []byte, deliveryID string) {
-	cfg := s.loadCfg()
-	var payload struct {
-		Action      string             `json:"action"`
-		Comment     webhookComment     `json:"comment"`
-		PullRequest webhookPullRequest `json:"pull_request"`
-		Repository  webhookRepository  `json:"repository"`
-		Sender      webhookSender      `json:"sender"`
-	}
-	if err := json.Unmarshal(body, &payload); err != nil {
-		http.Error(w, "invalid payload", http.StatusBadRequest)
-		return
-	}
-	if payload.Action != "created" {
-		w.WriteHeader(http.StatusAccepted)
-		return
-	}
-
-	repo, ok := cfg.RepoByName(payload.Repository.FullName)
-	if !ok || !repo.Enabled {
-		w.WriteHeader(http.StatusAccepted)
-		return
-	}
-
-	ev := workflow.Event{
-		ID:     deliveryID,
-		Repo:   workflow.RepoRef{FullName: repo.Name, Enabled: repo.Enabled},
-		Kind:   "pull_request_review_comment.created",
-		Number: payload.PullRequest.Number,
-		Actor:  payload.Sender.Login,
-		Payload: map[string]any{
-			"body": payload.Comment.Body,
-		},
-	}
-	s.enqueue(ctx, w, ev, deliveryID)
-}
-
-// handlePushEvent handles X-GitHub-Event: push.
-func (s *Server) handlePushEvent(ctx context.Context, w http.ResponseWriter, body []byte, deliveryID string) {
-	cfg := s.loadCfg()
-	var payload struct {
-		Ref        string            `json:"ref"`
-		After      string            `json:"after"`
-		Repository webhookRepository `json:"repository"`
-		Sender     webhookSender     `json:"sender"`
-	}
-	if err := json.Unmarshal(body, &payload); err != nil {
-		http.Error(w, "invalid payload", http.StatusBadRequest)
-		return
-	}
-
-	repo, ok := cfg.RepoByName(payload.Repository.FullName)
-	if !ok || !repo.Enabled {
-		w.WriteHeader(http.StatusAccepted)
-		return
-	}
-
-	// Ignore branch deletions (After is all-zero SHA) and non-branch refs
-	// (tags, notes). Only "new commit pushed to a branch" maps to push events.
-	const deletedSHA = "0000000000000000000000000000000000000000"
-	if payload.After == deletedSHA || !strings.HasPrefix(payload.Ref, "refs/heads/") {
-		w.WriteHeader(http.StatusAccepted)
-		return
-	}
-
-	ev := workflow.Event{
-		ID:    deliveryID,
-		Repo:  workflow.RepoRef{FullName: repo.Name, Enabled: repo.Enabled},
-		Kind:  "push",
-		Actor: payload.Sender.Login,
-		Payload: map[string]any{
-			"ref":      payload.Ref,
-			"head_sha": payload.After,
-		},
-	}
-	s.enqueue(ctx, w, ev, deliveryID)
-}
-
-// enqueue pushes ev onto the event queue, handling all error cases.
-func (s *Server) enqueue(ctx context.Context, w http.ResponseWriter, ev workflow.Event, deliveryID string) {
-	if err := s.channels.PushEvent(ctx, ev); err != nil {
-		if errors.Is(err, workflow.ErrEventQueueFull) {
-			s.delivery.Delete(deliveryID)
-			s.logger.Warn().Str("repo", ev.Repo.FullName).Str("kind", ev.Kind).Msg("event queue full, dropping webhook")
-			http.Error(w, "event queue full, retry later", http.StatusServiceUnavailable)
-			return
-		}
-		if errors.Is(err, workflow.ErrQueueClosed) {
-			s.logger.Warn().Str("repo", ev.Repo.FullName).Msg("queue closed during shutdown, dropping webhook")
-			http.Error(w, "shutting down, retry later", http.StatusServiceUnavailable)
-			return
-		}
-		s.delivery.Delete(deliveryID)
-		http.Error(w, "request cancelled", http.StatusRequestTimeout)
-		return
-	}
-	w.WriteHeader(http.StatusAccepted)
-}
-
-// verifySignature checks the HMAC-SHA256 signature from X-Hub-Signature-256.
-// hmac.Equal is used for the final comparison to avoid timing attacks that
-// could leak information about the expected value through execution time.
-func verifySignature(payload []byte, secret, signature string) bool {
-	signature = strings.TrimPrefix(strings.TrimSpace(signature), "sha256=")
-	if signature == "" || secret == "" {
-		return false
-	}
-	mac := hmac.New(sha256.New, []byte(secret))
-	_, _ = mac.Write(payload)
-	expected := hex.EncodeToString(mac.Sum(nil))
-	return hmac.Equal([]byte(expected), []byte(signature))
-}
